backend/requirements.txt
fastapi 
uvicorn 
python-multipart 
assemblyai
dotenv
google-genai
langchain-google-genai
faiss-cpu
numpy
pandas
requests
tavily-python
langchain-tavily
gTTS
aiortc  
pipecat-ai[daily,silero,google,openai]
